+++ widget/src/_standalone/tidy_feedback/embed.ts
@@ -1,0 +1,5 @@
+import { autoEmbedWithTarget } from 'svelte-standalone';
+
+import Tidy_feedback from './index.svelte';
+
+autoEmbedWithTarget(Tidy_feedback, 'tidy_feedback');